weather_app/requirements.txt
sense-hat==2.6.0
luma.oled==3.13.0
requests==2.31.0
numpy==1.24.3
smbus2==0.4.2
pandas==2.0.3
matplotlib==3.7.2
schedule==1.2.0
